review: split OIDC
